+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -2,7 +2,7 @@ object Dependencies {
     const val material = "com.google.android.material:material:1.4.0-alpha01"
 
     object ComposeMaterialDialogs {
-        const val version = "0.3.1"
+        const val version = "0.3.2"
 
         const val core = "com.vanpra.compose-material-dialogs:core:$version"
         const val datetime = "com.vanpra.compose-material-dialogs:datetime:$version"
@@ -22,10 +22,10 @@ object Dependencies {
     object AndroidX {
         const val appcompat = "androidx.appcompat:appcompat:1.2.0-rc01"
         const val coreKtx = "androidx.core:core-ktx:1.5.0-alpha03"
-        const val composeActivity = "androidx.activity:activity-compose:1.3.0-alpha04"
+        const val composeActivity = "androidx.activity:activity-compose:1.3.0-alpha05"
 
         object Compose {
-            const val version = "1.0.0-beta02"
+            const val version = "1.0.0-beta03"
 
             const val ui = "androidx.compose.ui:ui:$version"
             const val material = "androidx.compose.material:material:$version"