+++ datetime/src/main/java/com/vanpra/composematerialdialogs/datetime/DateTimePicker.kt
@@ -14,10 +14,14 @@ import kotlinx.coroutines.launch
 import java.time.LocalDateTime
 
 /**
- * @brief A date time picker dialog
+ * @brief A combined date and time picker dialog
  *
  * @param initialDateTime The date and time to be shown to the user when the dialog is first shown.
  * Defaults to the current date and time if this is not set
+ * @param timePickerColors see [TimePickerColors]
+ * @param yearRange the range of years the user should be allowed to pick from
+ * @param positiveButtonText text used for positive button label
+ * @param negativeButtonText text used for negative button label
  * @param onComplete callback with a LocalDateTime object when the user completes their input
  * @param onCancel callback when the user cancels the dialog
  */
@@ -26,6 +30,8 @@ fun MaterialDialog.datetimepicker(
     initialDateTime: LocalDateTime = LocalDateTime.now(),
     timePickerColors: TimePickerColors = TimePickerDefaults.colors(),
     yearRange: IntRange = IntRange(1900, 2100),
+    positiveButtonText: String = "Ok",
+    negativeButtonText: String = "Cancel",
     onCancel: () -> Unit = {},
     onComplete: (LocalDateTime) -> Unit = {}
 ) {
@@ -76,7 +82,7 @@ fun MaterialDialog.datetimepicker(
             text = if (isDateScreen) {
                 "Next"
             } else {
-                "Ok"
+                positiveButtonText
             },
             disableDismiss = isDateScreen
         ) {
@@ -94,7 +100,7 @@ fun MaterialDialog.datetimepicker(
             }
         }
 
-        negativeButton("Cancel") {
+        negativeButton(negativeButtonText) {
             onCancel()
         }
     }